Финальные правки для задания Надо подкачаться
