+++ js/validator.js
@@ -1,10 +1,9 @@
 import {sendData} from './api.js';
 import {closePhotoModal} from './processing-modal.js';
+import {isEscapeKey} from './util.js';
 
 const form = document.querySelector('.img-upload__form');
 const submitButton = document.querySelector('.img-upload__submit');
-const successTemplate = document.querySelector('#success').content;
-const errorTemplate = document.querySelector('#error').content;
 
 const pristine = new Pristine(form, {
   classTo: 'img-upload__text',
@@ -55,6 +54,37 @@ const initUploadWindow = () => {
     'Используйте только буквы и цифры'
   );
 
+  const successMessageTemplate = document.querySelector('#success').content.querySelector('.success');
+  const errorMessageTemplate = document.querySelector('#error').content.querySelector('.error');
+
+  const onEscKeydown = (evt) => {
+    if (isEscapeKey(evt)) {
+      evt.preventDefault();
+      closePhotoModal();
+      document.removeEventListener('keydown', onEscKeydown);
+    }
+  };
+
+  const onSucceed = () => {
+    const message = successMessageTemplate.cloneNode(true);
+    message.querySelector('.success__button').addEventListener('click', () => {
+      message.remove();
+      closePhotoModal();
+    });
+    document.addEventListener('keydown', onEscKeydown);
+    document.body.appendChild(message);
+  };
+
+  const onFailed = () => {
+    const message = errorMessageTemplate.cloneNode(true);
+    message.querySelector('.error__button').addEventListener('click', () => {
+      message.remove();
+      closePhotoModal();
+    });
+    document.addEventListener('keydown', onEscKeydown);
+    document.body.appendChild(message);
+  };
+
   form.addEventListener('submit', (evt) => {
     evt.preventDefault();
     evt.stopPropagation();
@@ -65,14 +95,12 @@ const initUploadWindow = () => {
         () => {
           closePhotoModal();
           unblockSubmitButton();
-          const message = successTemplate.cloneNode(true);
-          document.body.appendChild(message);
+          onSucceed();
         },
         () => {
           closePhotoModal();
           unblockSubmitButton();
-          const message = errorTemplate.cloneNode(true);
-          document.body.appendChild(message);
+          onFailed();
         },
         new FormData(evt.target),
       );
@@ -80,4 +108,4 @@ const initUploadWindow = () => {
   });
 };
 
-export {initUploadWindow};
+export {initUploadWindow, unblockSubmitButton};
